requirements.txt
asgiref==3.9.1
azure-core==1.35.1
azure-storage-blob==12.26.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
Django==5.2.6
gunicorn==23.0.0
idna==3.10
isodate==0.7.2
packaging==25.0
pycparser==2.23
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0